requirements.txt
Flask==2.3.3
fastapi==0.104.1
pydantic==2.4.1
Requests==2.31.0
SQLAlchemy==2.0.20
python-dotenv
PyMySQL
mysql-connector-python
alembic